fix tour flow seed + data type

+++ Data/Entities/TourFlow.cs
@@ -8,11 +8,11 @@ public class TourFlow
 
     public virtual Tour Tour { get; set; } = null!;
 
-    public float Longitude { get; set; }
+    public double Longitude { get; set; }
 
-    public float Latitude { get; set; }
+    public double Latitude { get; set; }
 
     public DateTime ArrivalTime { get; set; }
-    
+
     public string? Description { get; set; } = null!;
 }
